@@ -11,8 +11,8 @@
     </PropertyGroup>
     
     <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <LangVersion>8</LangVersion>
+        <TargetFramework>net8.0</TargetFramework>
+        <LangVersion>10.0</LangVersion>
         <Nullable>enable</Nullable>
         <WarningsAsErrors>CS8653;CS8625;CS8623;CS8618;CS8603;CS8602;CS8600</WarningsAsErrors>
     </PropertyGroup>
